+++ helpers/eventwebhook/eventwebhook_test.go
@@ -1,4 +1,4 @@
-package securewebhook
+package eventwebhook
 
 import (
 	"encoding/json"
@@ -29,15 +29,15 @@ func TestSecureWebhookNewSettings(t *testing.T) {
 	assert.NotNil(t, NewSettings(), "NewSettings() shouldn't return nil")
 }
 
-func TestSecureWebhookSetEnable(t *testing.T) {
+func TestSetSecureWebhookEnable(t *testing.T) {
 	s := NewSettings()
 	assert.NotNil(t, NewSettings(), "NewSettings() shouldn't return nil")
 
-	s.SetEnable(true)
-	assert.Equal(t, true, *s.Enable, fmt.Sprintf("SecureWebhook.Enable should be 'true', got %v", *s.Enable))
+	s.SetSecureWebhookEnable(true)
+	assert.Equal(t, true, *s.SecureWebhookEnable, fmt.Sprintf("SecureWebhook.Enable should be 'true', got %v", *s.SecureWebhookEnable))
 
-	s.SetEnable(false)
-	assert.Equal(t, false, *s.Enable, fmt.Sprintf("SecureWebhook.Enable should be 'false', got %v", *s.Enable))
+	s.SetSecureWebhookEnable(false)
+	assert.Equal(t, false, *s.SecureWebhookEnable, fmt.Sprintf("SecureWebhook.Enable should be 'false', got %v", *s.SecureWebhookEnable))
 }
 
 func TestSecureWebhookGetRequestBody(t *testing.T) {
@@ -47,12 +47,12 @@ func TestSecureWebhookGetRequestBody(t *testing.T) {
 	s := NewSettings()
 	assert.NotNil(t, NewSettings(), "NewSettings() shouldn't return nil")
 
-	s.SetEnable(false)
+	s.SetSecureWebhookEnable(false)
 	actualJSON, err := GetRequestBody(s)
 	require.NoError(t, err)
 	assert.Equal(t, expectedJSONDisabled, actualJSON, fmt.Sprintf("SecureWebhook.Enable should be '%b', got %b", expectedJSONDisabled, actualJSON))
 
-	s.SetEnable(true)
+	s.SetSecureWebhookEnable(true)
 	actualJSON, err = GetRequestBody(s)
 	require.NoError(t, err)
 	assert.Equal(t, expectedJSONEnabled, actualJSON, fmt.Sprintf("SecureWebhook.Enable should be '%b', got %b", expectedJSONEnabled, actualJSON))
